fix(teardown): release physical worktree + drop usage_limit_notify on delete (#1906)
Two TIER-1 teardown leaks in full_delete_instance (dev-2 audit, RCA verified),
same stale-state-on-redeploy class as #1902/#1903.

Leak 1 (physical worktree) — binding::unbind cleared binding.json but never
released the physical `worktrees/<name>/<branch>/`; the audit only saw
runtime/<name>/binding.json → false-clean; a same-name+branch redeploy then
collided with the orphan worktree (lease conflict). #1885 GC only reclaimed it a
later tick (backstop, not the intended clean path — full_delete forgetting it WAS
the leak). Fix (spike-approved Option a, eager release is safe — the agent is
already dead via delete_transaction → wait_for_child_exit):
  - full_delete calls `release_full(home, name, false)` (reuses the vetted
    clean-release path: marker-checked remove_worktree, deletes MERGED branches /
    PRESERVES UNMERGED — `false` is dry_run, not force). ORDER IS LOAD-BEARING:
    BEFORE binding::unbind, since release_full reads binding::read to locate the
    worktree. Then `remove_dir` (empty-only) drops the now-empty agent dir.
  - name_residual_anywhere now detects `worktrees/<name>/`.

Leak 2 (usage_limit_notify) — full_delete had ZERO cleanup of the instance's name
entry in usage_limit_notify.json, so a same-name redeploy inherited stale
suppression and ate its first real usage_limit notify. Same class as
escalation_persist::remove (#1680) but this store was missed. Fix:
  - supervisor::remove_usage_limit_notify (locked RMW, mirrors
    escalation_persist::remove) called in full_delete; + usage_limit_notify_has
    for the audit.

§3.9 (isolated home, never touches ~/.agend-terminal): audit detects each
residual; single delete removes the real uuid... err, the usage entry + releases
the physical worktree (worktrees/<name>/ gone).

File split: full_delete + audit additions pushed instance_lifecycle.rs over the
file_size_invariant 750-LOC cap; extracted the test module to
instance_lifecycle/tests.rs (subdir, per the invariant's split-don't-bypass
intent + the convention that dedicated tests.rs files are size/metadata-resolver
exempt). impl 322 LOC, tests 551 LOC — both under cap.

Full suite 3791/3791 under nextest --profile ci; fmt + clippy -D clean.


Agend-Agent: fixup-dev
Agend-Branch: fix/1906-teardown-leaks
Agend-Issued-At: 2026-06-09T09:08:51.691138+00:00
